Move SemVer into `Versioning` namespace
