+++ src/Program.cs
@@ -5,6 +5,7 @@ using Microsoft.Extensions.CommandLineUtils;
 using Skarp.Version.Cli.CsProj;
 using Skarp.Version.Cli.CsProj.FileSystem;
 using Skarp.Version.Cli.Vcs.Git;
+using Skarp.Version.Cli.Versioning;
 
 namespace Skarp.Version.Cli
 {
